diagnostico_project/requirements.txt
amqp==2.4.2
Babel==2.6.0
billiard==3.6.0.0
celery==4.3.0
Django==2.1.7
django-appconf==1.0.3
django-celery-email==2.0.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-phonenumber-field==2.3.0
django-webpack-loader==0.6.0
gunicorn==19.9.0
kombu==4.5.0
phonenumbers==8.10.8
Pillow==5.4.1
psycopg2-binary==2.7.7
pydash==4.7.4
pydjax==0.2
python-dotenv==0.10.1
pytz==2019.1
rc-django-celery-results==1.0.1
roman==3.2
six==1.12.0
sqlparse==0.3.0
vine==1.3.0
whitenoise==4.1.2
